refactor(find-funders): extract entity list into its own component
The grouping change pushed chat-view-client.tsx past the 600-line
quality-gate limit. Move the entity card / group rendering into
entity-list.tsx (606 → 359 lines); no behavior change.
